Split `CommandEnv` into OS specific implementations
Also add unsupported implementation of `CommandEnv`.
