+++ library/std/src/sys_common/mod.rs
@@ -26,7 +26,6 @@ pub mod fs;
 pub mod io;
 pub mod memchr;
 pub mod mutex;
-pub mod process;
 pub mod remutex;
 #[macro_use]
 pub mod rt;
